+++ src-tauri/src/main.rs
@@ -1,4 +1,4 @@
-// Prevents additional console window on Windows in release, DO NOT REMOVE!!
+// Windows 发布版下不要额外弹出控制台窗口，这行不能删。
 #![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
 
 fn main() {